+++ example/test.dart
@@ -1,0 +1,1 @@
+import 'package:objd/core.dart';